+++ src/components/MenuAction.js
@@ -26,7 +26,6 @@ export class MenuAction extends LitElement {
         a:active,
         a:focus {
           background-color: var(--pzsh-menu-bg-alt);
-          text-decoration: none;
         }
         ::slotted(pzsh-icon),
         ::slotted(svg) {